+++ src/test/groovy/io/github/devatherock/ldapsearch/controllers/AdditionalEndpointsTestSpec.groovy
@@ -1,0 +1,10 @@
+package io.github.devatherock.ldapsearch.controllers
+
+import io.micronaut.test.extensions.spock.annotation.MicronautTest
+
+/**
+ * Unit test for additional endpoints
+ */
+@MicronautTest
+class AdditionalEndpointsTestSpec extends AdditionalEndpointsSpec {
+}